refactor: rename kind config and move the configurator to kind module
