+++ annotations/kind-annotations/src/main/java/io/dekorate/kind/annotation/KindApplication.java
@@ -26,9 +26,19 @@ import io.sundr.builder.annotations.Buildable;
 import io.sundr.builder.annotations.Pojo;
 
 @Buildable(builderPackage = "io.fabric8.kubernetes.api.builder")
-@Pojo(name = "KindConfig", relativePath = "../config", autobox = true, mutable = true, superClass = BaseConfig.class, withStaticBuilderMethod = true, withStaticAdapterMethod = false, adapter = @Adapter(suffix = "Adapter", relativePath = "../adapter", withMapAdapterMethod = true))
+@Pojo(name = "KindLoadConfig", relativePath = "../config", autobox = true, mutable = true, superClass = BaseConfig.class, withStaticBuilderMethod = true, withStaticAdapterMethod = false, adapter = @Adapter(name = "KindLoadConfigAdapter", relativePath = "../adapter", withMapAdapterMethod = true))
 @Target({ ElementType.CONSTRUCTOR, ElementType.TYPE })
 @Retention(RetentionPolicy.RUNTIME)
 public @interface KindApplication {
+
+  boolean enabled() default true;
+
+  /**
+   * Flag to trigger the registration of the deploy hook. It's generally
+   * preferable to use `-Ddekorate.deploy=true` instead of hardcoding this here.
+   *
+   * @return True for automatic registration of the build hook.
+   */
+  boolean autoDeployEnabled() default true;
 
 }
