+++ annotations/kind-annotations/src/main/java/io/dekorate/kind/decorator/ApplyKindImageAutoloadConfiguration.java
@@ -1,4 +1,4 @@
-package io.dekorate.kind.configurator;
+package io.dekorate.kind.decorator;
 
 import io.dekorate.kubernetes.config.BaseConfigFluent;
 import io.dekorate.kubernetes.config.Configurator;
@@ -10,6 +10,6 @@ public class ApplyKindImageAutoloadConfiguration extends Configurator<BaseConfig
   @Override
   public void visit(BaseConfigFluent config) {
     config.withAutoLoadEnabled(
-        Boolean.parseBoolean(System.getProperty(DEKORATE_KIND_AUTOLOAD, String.valueOf(config.getAutoLoadEnabled()))));
+        Boolean.parseBoolean(System.getProperty(DEKORATE_KIND_AUTOLOAD, "true")));
   }
 }
